requirements.txt
# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Data
numpy>=1.24.0
pillow>=9.5.0
albumentations>=1.3.0

# ML Tools
scikit-learn>=1.3.0

# Config & Logging
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
tqdm>=4.65.0

# Optional: Monitoring
# tensorboard>=2.13.0
# wandb>=0.15.0
